requirements.txt
astroid==2.9.3
black==22.1.0
click==8.0.4
Flask==2.0.3
isort==5.10.1
itsdangerous==2.1.1
Jinja2==3.0.3
lazy-object-proxy==1.7.1
MarkupSafe==2.1.1
mccabe==0.6.1
mypy-extensions==0.4.3
paho-mqtt==1.6.1
pathspec==0.9.0
platformdirs==2.5.1
pylint==2.12.2
toml==0.10.2
tomli==2.0.1
typing_extensions==4.1.1
Werkzeug==2.0.3
wrapt==1.13.3
